pyproject.toml
[project]
name = "cfb-ratings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jacob Baldwin", email = "baldwinj30@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pandera[polars]>=0.22.1",
    "polars>=1.18.0",
    "requests>=2.32.3",
]

[project.scripts]
cfb-ratings = "cfb_ratings:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=8.31.0",
    "mypy>=1.14.1",
    "pre-commit>=4.0.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.4",
    "types-requests>=2.32.0.20241016",
]

[tool.ruff]
exclude = [
    ".git",
    ".venv",
    ".vscode",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "__pycache__",
]
# exclude files that match the list, even if explicitly passed via CLI;
# this forces pre-commit to honor ruff's exclude list
force-exclude = true
# run auto fixes, by default just for imports
fix = true

[tool.ruff.lint]
# add pydocstyle and isort rules to flake8
extend-select = ["D", "I"]
# auto-fix imports
fixable = ["I"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pycodestyle]
# this should generally let the formatter and linter agree, since
# the formatter has 88 as a soft limit
max-line-length = 99
